Commit 4 channel dewarp code
Closely based on the 16 channel code, but packs data differently.  FIR filtering is only done 256 bits at a time (which could be improved) by packing 4 pairs of 4 channels together rather than 2 pairs of 16, so it is slower than it could be.  Hermite interpolation also has additional permutes which is slightly slower, although it proceeds 256 bits at a time like the 16 channel case.